Scripts for refreshing swagger json

+++ tools/swagger_gen.sh
@@ -1,12 +1,16 @@
 #!/bin/bash
 
-set -ex
+set -e
 
 echo "Clear ochk/sdk/gen directory"
 rm -rf ochk/sdk/gen/*
 
-echo "Install swagger"
-go install github.com/go-swagger/go-swagger/cmd/swagger
+SWAGGER_BIN="$(go env GOPATH)/bin/swagger"
+
+if [[ ! -f "${SWAGGER}" ]]; then
+  echo "Install swagger"
+  go install github.com/go-swagger/go-swagger/cmd/swagger
+fi
 
 echo "Run swagger gen"
-"$(go env GOPATH)/bin/swagger" generate client -f ./ochk/sdk/swagger.json -t ./ochk/sdk/gen/ -A ochk
+${SWAGGER_BIN} generate client -f ./ochk/sdk/swagger.json -t ./ochk/sdk/gen/ -A ochk
